chore(StatusOverTime): move styled components into separate file
